Centralize iframe constants in lib/iframe-utils
Address ce:review findings:
- Move UNSUPPORTED_IFRAME_SRCS and frame-depth clamping into
  lib/iframe-utils.js so the constants don't drift across SDKs.
- isUnsupportedIframeSrc now lowercases the src so JavaScript:/DATA:
  prefixes are caught.
- resolveMaxFrameDepth/resolveIgnoreSelectors now compose clampFrameDepth
  and normalizeIgnoreSelectors from the shared utils module.
