+++ lib/snapshot.js
@@ -1,24 +1,12 @@
+const {
+  DEFAULT_MAX_FRAME_DEPTH,
+  isUnsupportedIframeSrc,
+  clampFrameDepth,
+  normalizeIgnoreSelectors
+} = require('./iframe-utils');
+
 const CS_MAX_SCREENSHOT_LIMIT = 25000;
 const SCROLL_DEFAULT_SLEEP_TIME = 0.45; // 450ms
-const DEFAULT_MAX_FRAME_DEPTH = 10;
-// Hard ceiling so a misconfigured option can't trigger pathological recursion
-const HARD_MAX_FRAME_DEPTH = 25;
-
-const UNSUPPORTED_IFRAME_SRCS = [
-  'about:blank',
-  'about:srcdoc',
-  'javascript:',
-  'data:',
-  'blob:',
-  'vbscript:',
-  'chrome:',
-  'chrome-extension:'
-];
-
-function isUnsupportedIframeSrc(src) {
-  if (!src) return true;
-  return UNSUPPORTED_IFRAME_SRCS.some(prefix => src === prefix || src.startsWith(prefix));
-}
 
 function getOrigin(url) {
   try {
@@ -85,19 +73,19 @@ function getCookies(browser) {
 }
 
 function resolveMaxFrameDepth(options = {}, utils) {
-  const raw = options.maxIframeDepth
-    ?? utils?.percy?.config?.snapshot?.maxIframeDepth
-    ?? DEFAULT_MAX_FRAME_DEPTH;
-  const n = Number(raw);
-  if (!Number.isFinite(n) || n < 1) return DEFAULT_MAX_FRAME_DEPTH;
-  return Math.min(n, HARD_MAX_FRAME_DEPTH);
+  return clampFrameDepth(
+    options.maxIframeDepth
+      ?? utils?.percy?.config?.snapshot?.maxIframeDepth
+      ?? DEFAULT_MAX_FRAME_DEPTH
+  );
 }
 
 function resolveIgnoreSelectors(options = {}, utils) {
-  const list = options.ignoreIframeSelectors
-    ?? utils?.percy?.config?.snapshot?.ignoreIframeSelectors
-    ?? [];
-  return Array.isArray(list) ? list.filter(s => typeof s === 'string' && s.trim()) : [];
+  return normalizeIgnoreSelectors(
+    options.ignoreIframeSelectors
+      ?? utils?.percy?.config?.snapshot?.ignoreIframeSelectors
+      ?? []
+  );
 }
 
 async function enumerateIframesInCurrentContext(browser, ignoreSelectors = []) {
